For Python, explicitly convert int64 to int32 which is what GCAM will be using for years and period values and makes us robust to what numpy may choose to use as default.  This closes #6

+++ inst/include/interp_interface.h
@@ -193,7 +193,7 @@ namespace Interp {
     using DataFrame = bp::dict;
     using NumericVector = NumpyVecWrapper<double>;
     using StringVector = NumpyVecWrapper<bp::str>;
-    using IntegerVector = NumpyVecWrapper<size_t>;
+    using IntegerVector = NumpyVecWrapper<int>;
     using NumericMatrix = bnp::ndarray;
 
     inline DataFrame createDataFrame() {
